--- src/coord/src/session/transaction.rs
@@ -1,19 +1,0 @@
-// Copyright Materialize, Inc. All rights reserved.
-//
-// Use of this software is governed by the Business Source License
-// included in the LICENSE file.
-//
-// As of the Change Date specified in that file, in accordance with
-// the Business Source License, use of this software will be governed
-// by the Apache License, Version 2.0.
-
-/// The current transaction status of a Session
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum TransactionStatus {
-    /// Not currently in a transaction
-    Idle,
-    /// Currently in a transaction
-    InTransaction,
-    /// Currently in a transaction block which is failed
-    Failed,
-}
